Port server traits and types from siminau-rpc
Since siminau-rpc is intended to hold rpc message types only, it is out
of scope for that project to also include types that allow for setting
up a general siminau server. As such, those out-of-scope types are being
moved to this new lyra crate.

Please note: the traits and types and integration test was moved as is.
This means that currently there are quite a few dev dependencies that
might not necessarily be needed. It is expected that this project will
be refactored, and some of the current dev dependencies not needed
anymore.