codec: port MsgPackCodec from siminau-rpc crate
Since the siminau-rpc crate implements serialization and deserialization
of messages via the ``AsBytes`` and ``FromBytes`` traits, siminau-rpc's
codec module is now obsolete and will be removed from that crate in the
near future. However, in order for the integration test to successfully
run, ``MsgPackCodec`` is still needed.

This changeset thus implements MsgPackCodec in this crate by using the
message serialization and deserialization traits defined in the
siminau-rpc crate.

Note: This implementation is meant to be a port only, no further
enhancements were developed.